\name{pm25}
\docType{data}
\title{PM2.5 values in Southern US}
\description{
  This data set contains the pm2.5 values in Southern US for the year 2010 and 
  multiple variates.
}
\format{A data frame of 5000 rows * 16 columns.}
\references{
  Hu X, Waller LA, Lyapustin A, Wang Y, Al-Hamdan MZ, Crosson WL, Estes Jr MG, Estes SM, Quattrochi DA, Puttaswamy SJ, Liu Y. Estimating ground-level PM2. 5 concentrations in the Southeastern United States using MAIAC AOD retrievals and a two-stage model. Remote Sensing of Environment. 2014 Jan 1;140:220-32
}
\keyword{datasets}