% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.wbart_sp.R
\name{predict.wbart_sp}
\alias{predict.wbart_sp}
\title{Predicting new observations with a previously fitted wbart_sp model}
\usage{
\method{predict}{wbart_sp}(
  object,
  newdata,
  newloc,
  draw_from_total_distribution,
  block = 1,
  seed = 88,
  ...
)
}
\arguments{
\item{object}{An \code{object} of type \code{wbart_sp} (fitted from \code{wbart_sp} function).}

\item{newdata}{A data frame of covariates to predict for.}

\item{newloc}{A data frame of location information of the obervations in \code{newdata}.
It should have 2 columns (names should be exactly the same as those of the \code{coordinates_train} when fitting the model).}

\item{draw_from_total_distribution}{Whether sampling from the total distribution when doing prediction on \code{newdata}. If so, it would be slower but utilize all the location information.}

\item{block}{If \code{draw_from_total_distribution=FALSE}, how many locations should be drawn at the same time.}

\item{seed}{Setting the seed for reproducibility.}
}
\value{
The return is a list containing these components:

\code{fhat.test} a matrix of drawings of \eqn{f} corresponding to newdata. Each row corresponds to a draw of the spatial random effect and each column corresponds to a row of \code{newdata}

\code{yhat.test} a matrix of final predictions (sum of \code{fhat.test} and \code{what.test}) corresponding to newdata. Each row corresponds to a draw of the spatial random effect and each column corresponds to a row of \code{newdata}

\code{what.test} a matrix of drawings of spatial random effect corresponding to newdata. Each row corresponds to a draw of the spatial random effect and each column corresponds to a row of \code{newdata}.

\code{unique_test_locations} A data frame of unique test locations in
\code{newdata} (order not necessary the same as in \code{newdata}).
}
\description{
This function is used to predict outcomes of new observations with a previously fitted model with type \code{wbart_sp}.
}
