% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbart_sp.R
\name{wbart_sp}
\alias{wbart_sp}
\title{Spatially Adjusted Bayesian Additive Regression Trees for continuoues outcomes}
\usage{
wbart_sp(
  x_train,
  y_train,
  coordinates_train,
  coordinates_system,
  x_test = matrix(0, 0, 0, 0),
  coordinates_test = NULL,
  sparse = FALSE,
  theta = 0,
  omega = 1,
  a = 0.5,
  b = 1,
  rho = NULL,
  augment = FALSE,
  xinfo = matrix(0, 0, 0),
  usequants = FALSE,
  cont = FALSE,
  rm.const = TRUE,
  sigest = NA,
  sigdf = 3,
  sigquant = 0.9,
  k = 2,
  power = 2,
  base = 0.95,
  sigmaf = NA,
  lambda = NA,
  fmean = mean(y_train),
  w = rep(1, length(y_train)),
  ntree = 200L,
  numcut = 100L,
  ndpost = 1000L,
  nskip = 100L,
  keepevery = 1L,
  nkeeptrain = ndpost,
  nkeeptest = ndpost,
  nkeeptestmean = ndpost,
  nkeeptreedraws = ndpost,
  printevery = 100L,
  transposed = FALSE,
  logrange_select_sd = 0.3,
  logsmoothness_select_sd = 0.3,
  sigma2_prior_a = 10,
  sigma2_prior_b = 1,
  tau2_prior_a = 1,
  tau2_prior_b = 1,
  logrange_init = 0,
  logsmoothness_init = 0,
  tau2_init = 1,
  logrange_prior_mean = 1,
  logrange_prior_sd = 0.5,
  logsmoothness_prior_mean = 0,
  logsmoothness_prior_sd = 0.5,
  mc,
  mc.cores = 2,
  draw_from_total_distribution = TRUE,
  block = 50,
  seed = 88
)
}
\arguments{
\item{x_train}{Explanatory variables for training (in sample) data.
Must be a data frame, with rows corresponding to observations and columns to variables.
If a variable is a factor in a data frame, it is replaced with dummies. Note that q dummies are created if q>2 and one dummy is created if q=2, where q is the number of levels of the factor.
Location information can be either in \code{x_train} or not but must be in \code{coordinates_train}.}

\item{y_train}{Continuous dependent variable for training (in sample) data. Must be a vector whose length equals to the number of rows in \code{x_train}.}

\item{coordinates_train}{The location information of observations in \code{x_train}. Must be a dataframe of 2 columns and the same number of rows as \code{x_train}. If latitude and longitude are provided as location information, the 2 columns of \code{coordinates_train} must be named exactly 'lon' and 'lat' (order matter) and the argument \code{coordiantes} set as 'lonlat'.
If locations information on the ground are provided, the 2 columns must be named exactly 'x' and 'y' (order matters) and argument \code{coordiantes} set as 'ground'.
The distane between locations is calculated accordingly (see \code{coordinates_system}).}

\item{x_test}{Explanatory variables for test (out of sample) data. Should have same structure as \code{x_train} (Must be a data frame, with rows corresponding to observations and columns to variables). If provided, must also provide \code{coordinates_test}.}

\item{coordinates_test}{The location information of observations in \code{x_test}. Should have same structure as \code{coordinates_train}. It must be of the same kind of coordinate system as \code{coordinates_train} and named exactly the same as \code{coordinates_train} (order matters).}

\item{logrange_select_sd}{Spatial residual sampling parameter. logRange select SD in mcmc. (see ...)}

\item{logsmoothness_select_sd}{Spatial residual sampling parameter. logSmoothness select SD in mcmc. (see ...)}

\item{sigma2_prior_a}{Prior paramter for the random noise \eqn{e}. (see ...)}

\item{sigma2_prior_b}{Prior paramter for the random noise \eqn{e}. (see ...)}

\item{tau2_prior_a}{Prior paramter for the matern correlation function tau2 (see ...)}

\item{tau2_prior_b}{Prior paramter for the matern correlation function tau2 (see ...)}

\item{logrange_init}{Initial value for logrange in mcmc.}

\item{logsmoothness_init}{Initial value for logsmoothness in mcmc.}

\item{tau2_init}{Initial value for tau2 in mcmc.}

\item{logrange_prior_mean}{Prior paramter for the matern correlation function logrange mean (see ...)}

\item{logrange_prior_sd}{Prior paramter for the matern correlation function logrange sd (see ...)}

\item{logsmoothness_prior_mean}{Prior paramter for the matern correlation function logsmoothness mean (see ...)}

\item{logsmoothness_prior_sd}{Prior paramter for the matern correlation function logsmoothness sd (see ...)}

\item{mc}{Whether fitting the model in parallel. (which usually improves the model performance but requires multiple cores.) Please also set the number of threads in argument \code{mc.cores}.}

\item{mc.cores}{How many threads to use if fitting the model in parallel. If \code{mc=FALSE}, this argument does not matter. If \code{mc=TRUE}, how many threads to use.}

\item{draw_from_total_distribution}{If draw from total distribution or ? distribution in the prediction. If no \code{x_test}, does not matter. Usually it would be slower but perserving and utilizing all the location information in the testing dataset to set \code{draw_from_total_distribution=TRUE} instead of \code{FALSE}.}

\item{block}{The spatial random effect of how many locations to predict at one time if \code{draw_from_total_distribution=TRUE}. If \code{draw_from_total_distribution=FALSE}, \code{block} is not used.}

\item{coordiantes_system}{What the \code{coordinates_train} are. Must be either 'lonlat' or 'ground'. If \code{coordinates_train} is the longitude and latitude information, \code{coordinates_train} should be 'lonlat' and the distance is calculated using grand circle distance with unit km. If \code{coordinates_train} is the location information on the ground, \code{coordinates_train} should be 'ground' and the distance is calculated using Euclidean distance.}
}
\value{
\code{wbart_sp} returns an object of type \code{wbart_sp} which is a list. It has the following components:

\code{fhat.train} A matrix with ndpost rows and nrow(x_train) columns. Each row corresponds to a draw \eqn{f^*}{f*} from the posterior of \eqn{f} and each column corresponds to a row of x_train. The \eqn{(i,j)} value is \eqn{f^*(x)}{f*(x)} for the \eqn{i^{th}}{i\^th} kept draw of \eqn{f} and the \eqn{j^{th}}{j\^th} row of x.train. Burn-in is dropped. NOTICE: this is the not final prediction value, \code{yhat.train} is.

\code{fhat.test} Same as \code{fhat.train} but now the x's are the rows of the test data.

\code{yhat.train} A matrix with ndpost rows and nrow(x_train) columns. Each row corresponds to the final prediction (sum of a draw from \eqn{f(x)} and a draw of the spatial random effect) and each column corresponds to a row of x_train.

\code{yhat.test} Same as \code{yhat.train} but now the x's are the rows of the test data.

\code{what.train} A matrix with ndpost rows and nrow(x_train) columns. Each row corresponds to a draw of the spatial random effect and each column corresponds to a row of x_train.

\code{what.test} Same as \code{what.train} but now the x's are the rows of the test data.

\code{sigma} post burn in draws of sigma, length = ndpost.

\code{sigma_all} A data frame of burn in draws and post burn in draws of sigma, dim = (nskip + ndpost/mc.cores) * (\code{mc.cores}). Can be used to inspect convergence.

\code{tau2} post burn in draws of sigma, length = ndpost.

\code{logrange} post burn in draws of logrange, length = ndpost.

\code{logsmoothness} post burn in draws of logsmoothness, length = ndpost.

\code{nskip} nskip

\code{ndpost} ndpost

\code{mu} mean of y_train

\code{varcount} a matrix with ndpost rows and nrow(x_train) columns. Each row is for a draw. For each variable (corresponding to the columns), the total count of the number of times that variable is used in a tree decision rule (over all trees) is given.

\code{varprob} a matrix with ndpost rows and nrow(x_train) columns. Each row is for a draw. For each variable (corresponding to the columns), the probability (frequency / total frequency) that variable is used in a tree decision rule (over all trees) is given.

\code{sigest} The rough error standard deviation (\eqn{\sigma}{sigma}) used in the prior.

\code{coordinates_system} Coordiantes parameters for \code{coordinates_train}.

\code{unique_train_lcations} Unique locations in the training data.

\code{unique_w} sampled spatial random effects according for \code{unique_train_locations}.

\code{proc.time} processing time
}
\description{
\code{wbart_sp} is a Bayesian "sum-of-trees" model designed for spatial data. It is
built upon the original BART model (see ...) with an extra spatial random effect.

For a numeric continuous outcome \eqn{y}, we have \eqn{y = f(x) + e_s + e}, where \eqn{e_s} is the spatial random effect and \eqn{e ~ N(0, sigma^2)}. See ...
}
\details{
\code{wbart_sp} is the only function (besides S3 method \code{predict.wbart_sp}) provided by
this package.

\code{wbart_sp} implements the spatially adjusted Bayesian Additive Regression Trees (in single
 thread or multiple threads). S3 method \code{predict.wbart_sp} implements the prediction of a model of class wbart_sp.

The detailed information about the model please see: paper/github ...
}
\references{
Chipman, H., George, E., and McCulloch R. (2010)
Bayesian Additive Regression Trees.
\emph{The Annals of Applied Statistics}, \bold{4,1}, 266-298 <doi:10.1214/09-AOAS285>.

Chipman, H., George, E., and McCulloch R. (2006)
Bayesian Ensemble Learning.
Advances in Neural Information Processing Systems 19,
Scholkopf, Platt and Hoffman, Eds., MIT Press, Cambridge, MA, 265-272.

Friedman, J.H. (1991)
Multivariate adaptive regression splines.
\emph{The Annals of Statistics}, \bold{19}, 1--67.

Linero, A.R. (2018)
Bayesian regression trees for high dimensional prediction and variable
selection. \emph{JASA}, \bold{113}, 626--36.
}
